[refac] move logic to university store package
