--- src/main/java/org/hankki/hankkiserver/api/store/controller/request/StoreCreateRequest.java
@@ -1,7 +1,0 @@
-package org.hankki.hankkiserver.api.store.controller.request;
-
-public record StoreCreateRequest(
-        Long storeId,
-        Long universityId
-) {
-}
