+++ src/main/java/org/hankki/hankkiserver/api/universitystore/controller/request/UniviersityStoreCreateRequest.java
@@ -1,0 +1,7 @@
+package org.hankki.hankkiserver.api.universitystore.controller.request;
+
+public record UniviersityStoreCreateRequest(
+        Long storeId,
+        Long universityId
+) {
+}
